requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
playwright>=1.40.0
pydantic>=2.0.0
python-telegram-bot>=21.0
httpx>=0.25.0
